+++ js-client/src/client.ts
@@ -1,7 +1,9 @@
+import { Attributes, trace } from '@opentelemetry/api';
 import { proxyUrl, propagateSpan } from './internal.js';
 import type {
   ChronicleChatRequest,
   ChronicleChatResponse,
+  ChronicleRequestMetadata,
   ChronicleRequestOptions,
 } from './types.js';
 
@@ -18,17 +20,19 @@ export interface ChronicleClientOptions {
   defaults?: Partial<ChronicleRequestOptions>;
 }
 
-export type ChronicleClient = (
+export type ChronicleEventFn = (event: ChronicleEvent) => Promise<{ id: string }>;
+export type ChronicleClient = ((
   chat: ChronicleChatRequest & Partial<ChronicleRequestOptions>,
   options?: ChronicleRequestOptions
-) => Promise<ChronicleChatResponse>;
+) => Promise<ChronicleChatResponse>) & { event: ChronicleEventFn };
 
 /** Create a Chronicle proxy client. This returns a function which will call the Chronicle proxy */
 export function createChronicleClient(options?: ChronicleClientOptions): ChronicleClient {
   let { fetch = globalThis.fetch, token, defaults = {} } = options ?? {};
   let url = proxyUrl(options?.url);
+  let eventUrl = new URL('/event', url);
 
-  return async (
+  const client = async (
     chat: ChronicleChatRequest & Partial<ChronicleRequestOptions>,
     options?: ChronicleRequestOptions
   ) => {
@@ -64,26 +68,113 @@ export function createChronicleClient(options?: ChronicleClientOptions): Chronic
     if (res.ok) {
       return (await res.json()) as ChronicleChatResponse;
     } else {
-      let message = '';
-      const err = await res.text();
-      try {
-        const { error } = JSON.parse(err);
-
-        let errorBody = error?.details.body;
-        if (errorBody?.error) {
-          errorBody = errorBody.error;
-        }
-
-        if (errorBody) {
-          message = typeof errorBody === 'string' ? errorBody : JSON.stringify(errorBody);
-        }
-      } catch (e) {
-        message = err;
-      }
-
-      console.error(err);
-      // TODO The api returns a bunch of other error details, so integrate them here.
-      throw new Error(message || 'An error occurred');
+      throw new Error(await handleError(res));
     }
   };
+
+  client.event = (event: ChronicleEvent) => {
+    const thisEvent = {
+      ...event,
+      metadata: {
+        ...defaults.metadata,
+        ...event.metadata,
+      },
+      url: eventUrl,
+    };
+
+    return sendEvent(thisEvent);
+  };
+
+  return client;
+}
+
+export interface ChronicleEvent {
+  /** The type of event */
+  type: string;
+  /** Data associated with the event */
+  data?: object;
+  /** Error data for the event, if it represents an error. */
+  error?: any;
+  /** Additional metadata for the event */
+  metadata?: Omit<ChronicleRequestMetadata, 'extra'>;
+}
+
+export interface ChronicleSendEventOptions extends ChronicleEvent {
+  /** The URL to send the event to */
+  url?: string | URL;
+}
+
+export async function sendEvent(event: ChronicleSendEventOptions): Promise<{ id: string }> {
+  const { url, ...body } = event;
+  let req = new Request(proxyUrl(event.url, '/event'), {
+    method: 'POST',
+    headers: {
+      'content-type': 'application/json; charset=utf-8',
+    },
+    body: JSON.stringify(body),
+  });
+
+  const span = trace.getActiveSpan();
+  if (span?.isRecording) {
+    let eventAttributes: Attributes = {};
+
+    if (body.metadata) {
+      for (let k in body.metadata) {
+        eventAttributes[`llm.meta.${k}`] = body.metadata[k as keyof typeof body.metadata];
+      }
+    }
+
+    if (body.data) {
+      for (let k in body.data) {
+        let value = body.data[k as keyof typeof body.data];
+
+        let attrKey = `llm.meta.${k}`;
+        if (value && typeof value === 'object' && !Array.isArray(value)) {
+          eventAttributes[attrKey] = JSON.stringify(value);
+        } else {
+          eventAttributes[attrKey] = value;
+        }
+      }
+    }
+
+    if (body.error) {
+      eventAttributes['error'] =
+        typeof body.error === 'object' ? JSON.stringify(body.error) : body.error;
+    }
+
+    console.log(eventAttributes);
+
+    span.addEvent(body.type, eventAttributes);
+  }
+
+  let res = await fetch(req);
+  if (res.ok) {
+    const result = (await res.json()) as { id: string };
+    return result;
+  } else {
+    throw new Error(await handleError(res));
+  }
+}
+
+async function handleError(res: Response) {
+  let message = '';
+  const err = await res.text();
+  try {
+    const { error } = JSON.parse(err);
+
+    let errorBody = error?.details.body;
+    if (errorBody?.error) {
+      errorBody = errorBody.error;
+    }
+
+    if (errorBody) {
+      message = typeof errorBody === 'string' ? errorBody : JSON.stringify(errorBody);
+    }
+  } catch (e) {
+    message = err;
+  }
+
+  console.error(err);
+  // TODO The api returns a bunch of other error details, so integrate them here.
+  return message || 'An error occurred';
 }