+++ js-client/src/internal.ts
@@ -13,10 +13,10 @@ export function propagateSpan(req: Request) {
   propagator.inject(context.active(), req, setter);
 }
 
-export function proxyUrl(configured?: string) {
+export function proxyUrl(configured?: string | URL, path = '/chat') {
   let url = new URL(configured || process.env.CHRONICLE_PROXY_URL || 'http://localhost:9782');
   if (url.pathname.length <= 1) {
-    url.pathname = '/chat';
+    url.pathname = path;
   }
 
   return url;