+++ js-client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # chronicle-proxy JavaScript client changelog
 
+## Unreleased
+
+- Allow sending arbitrary events to the Chronicle proxy
+
 ## 0.1.0
 
 - Initial release